+++ test/test_HP3ParClient_CPG.py
@@ -1,5 +1,4 @@
-# vim: tabstop=4 shiftwidth=4 softtabstop=4
-# Copyright 2009-2012 10gen, Inc.
+# (c) Copyright 2015 Hewlett Packard Development Company, L.P.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.